+++ CHANGELOG.md
@@ -7,12 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+No changes
+
+## [1.3.0]
+
 ### Added
 
 * Make callout text into richtext, with a limited set of features; increase limit to 1000 chars
 * Support for specific per-page social-sharing images, at a reasonable size (#255)
-
-## [1.2.1]
 
 ### Changed